SelectViewWidget: Move it next to GameListHeadBar
Finally delete extra_widgets, finally!
